Update spi_logger.c
Revamped the debug section and added a way to enter a "debug mode" and order a dump of debug vars anytime a want.
Fixed a bug which broke the program when doing a maximum sample, ie. after a reset. The string was 1044 chars long and I only had 1000 chars allocated... Bad of me... :-(
Added error checking of mysql_options also, just in case...
Changed the init of the FIFO to make it more readable, but it still doesn't get the right permissions...