Update spi_logger.c
Necessary changes to match latest avrlogger changes to eliminate a race condition. We need to keep track of which sample area we are dealing with and respond accordingly. As a bonus we now can count the different types of samples and measure the time between each if we want to. For now we only time the TWI samples